requirements.txt
streamlit
opencv-python
keras
numpy
tensorflow
pillow
